requirements.txt
langchain==0.2.* 
langchain-community==0.2.* 
sentence-transformers==2.* 
chromadb==0.5.* 
pydantic==2.* 
pillow==10.* 
pytesseract==0.3.* 
opencv-python==4.* 
tqdm==4.* 
transformers
accelerate==0.31.* 
bitsandbytes==0.43.* 
peft==0.11.* 
trl==0.9.* 
datasets==2.* 
tokenizers
rich==13.* 
typer==0.12.* 
python-dotenv==1.* 
unidecode==1.* 
textstat==0.7.* 
nltk==3.*
pillow
numpy 
opencv-python 
pytesseract
matplotlib
faiss-cpu
